Componentized profile experiences
