+++ src/components/ProfessionalExperiences/styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+
+
+export const BodyAcademicForm = styled.section`
+  
+`;
